修改severcore

+++ server/servercore.cpp
@@ -22,9 +22,9 @@ servercore::servercore(TcpServer *ptcpserver, QObject *parent) : QObject(parent)
     connect(tp, &TcpServer::disconnected, this, &servercore::offline);
     db = QSqlDatabase::addDatabase("QODBC");
     db.setPort(3306);
-    db.setDatabaseName("oplinkbase");//不同电脑连接不同的数据库的时候记得改
-    //db.setUserName("root");
-    //db.setPassword("7979");
+    db.setDatabaseName("mysql");//不同电脑连接不同的数据库的时候记得改
+    db.setUserName("root");
+    db.setPassword("7979");
     db.open();
     if(!db.isOpen())qDebug()<<"Dont Connected"<<endl<<db.lastError().text()<<endl;
 }
@@ -85,8 +85,7 @@ void servercore::switchFunction(QTcpSocket *psocket)
 void servercore::returnSendTextMessageResult(bool Status, int MID, int SenderOID, int TargetOID, QString Type, QString Value)
 {
     // 还没实现文件消息的传输
-    QHostAddress targetip = socketmap.value(TargetOID)->peerAddress();
-    quint16 targetport = socketmap.value(TargetOID)->peerPort();
+
     QJsonObject returnJsonObject;
     returnJsonObject["Statues"]=Status;
     returnJsonObject["MID"]=MID;
@@ -97,6 +96,8 @@ void servercore::returnSendTextMessageResult(bool Status, int MID, int SenderOID
     returnJsonObject["Value"]=Value;
     if (Status)
     {
+        QHostAddress targetip = socketmap.value(TargetOID)->peerAddress();
+        quint16 targetport = socketmap.value(TargetOID)->peerPort();
         QJsonDocument returnJsonDocument(returnJsonObject);
         QByteArray returnJsonData = returnJsonDocument.toJson();
         qDebug()<<returnJsonData;
@@ -120,23 +121,32 @@ void servercore::SendTxtMessageRequest(QJsonObject &jsonObj)
     QString Value = jsonObj["Value"].toString();
 
     QSqlQuery query(db);
-    int MID = 0;
-    if (!query.exec(QString("SELECT MAX(%1) FROM message").arg(MID)))
+    int MID = -100;
+    if (!query.exec(QString("SELECT MAX(MID) AS MaxMID FROM message")))
     {
         qDebug() << "Query failed:" << query.lastError().text();
     }
     if (query.next()) {
-        MID = query.value(0).toInt();
+        MID = query.value("MaxMID").toInt();
+        qDebug()<<query.value("MaxMID");
         qDebug() << "Maximum value of" << "MID" << "is:" << MID++;
     }
     //判断消息类型是文本（Txt）还是文件（Doc），文本的Value为消息内容
+    qDebug()<<1;
+    QTcpSocket* flag = socketmap.value(TargetOID, nullptr);
+    qDebug()<<Type;
     if (Type == "Txt")
     {
-        if (socketmap.contains(TargetOID))
+        qDebug()<<123;
+
+        if (flag)
         {
+            qDebug()<<76513476;
             returnSendTextMessageResult(1, MID, SenderOID, TargetOID, Type, Value);
         } else
         {
+            qDebug()<<2;
+            qDebug()<<"MID : "<<MID;
             QDateTime SendTime = QDateTime::currentDateTime();
             query.prepare("INSERT INTO message (MID,Type,SenderOID,TargetOID,SendTime,Readed,Value) VALUES(:MID,:Type,:SenderOID,:TargetOID,:SendTime,:Readed,:Value)");
             query.bindValue(":MID", MID);
@@ -146,18 +156,20 @@ void servercore::SendTxtMessageRequest(QJsonObject &jsonObj)
             query.bindValue(":SendTime", SendTime);
             query.bindValue(":Readed", 0);
             query.bindValue(":Value", Value);
+            qDebug()<<66;
             if (!query.exec())
             {
                 qDebug() << "Database insertion error:" << query.lastError().text();
                 returnSendTextMessageResult(0, MID, SenderOID, TargetOID, Type, Value);
                 return;
             }
-            returnSendTextMessageResult(1, MID, SenderOID, TargetOID, Type, Value);
+            qDebug()<<3;
+            returnSendTextMessageResult(0, MID, SenderOID, TargetOID, Type, Value);
         }
     }
     else
     {
-        if (socketmap.contains(TargetOID))
+        if (flag)
         {
             returnSendTextMessageResult(1, MID, SenderOID, TargetOID, Type, Value);
         }
@@ -507,7 +519,7 @@ void servercore::FriendListRequest(QJsonObject &jsonObj)
 {
     int OID1 = jsonObj["OID"].toInt();
     QSqlQuery query(db);
-    if (!query.exec(QString("SELECT * FROM oplinkbase.friend join oplinkbase.account ON friend.OID1 =account.OID  WHERE OID1 = %1 ;").arg(OID1)))
+    if (!query.exec(QString("SELECT * FROM oplinkbase.friend join oplinkbase.account ON friend.OID2 =account.OID  WHERE OID1 = %1 ;").arg(OID1)))
     {
         qDebug() << "Query failed:" << query.lastError().text();
     }
@@ -540,6 +552,8 @@ void servercore::SynchronizeServerMessages(int OID)
             returnEnterResult(OID,sp->peerAddress(),0,"Query ERROR: "+query.lastError().text());
             return ;
         }
+        int Size = query.size();
+        qDebug()<<"Size : "<<Size;
         while(query.next())
         {
             QJsonObject returnJsonObject;
@@ -553,13 +567,17 @@ void servercore::SynchronizeServerMessages(int OID)
             qDebug()<<returnJsonData;
             tp->send(sp->peerAddress(), sp->peerPort(), returnJsonData);
         }
-        query.prepare("DELETE FROM message WHERE TargetOID = :OID");
-        query.bindValue(":OID", OID);
-        if (query.exec()) {
-            qDebug() << "Deletion successful";
-        } else {
-            qDebug() << "Deletion failed:" << query.lastError().text();
+        if(Size !=-1 )
+        {
+            query.prepare("DELETE FROM message WHERE TargetOID = :OID");
+            query.bindValue(":OID", OID);
+            if (query.exec()) {
+                qDebug() << "Deletion successful";
+            } else {
+                qDebug() << "Deletion failed:" << query.lastError().text();
+            }
         }
+
 }
 
 void servercore::GroupListRequest(QJsonObject &jsonObj)
